Service Registration

+++ OnionPronia/src/Core/OnionPronia.Domain/Entities/Product.cs
@@ -12,7 +12,7 @@ namespace OnionPronia.Domain.Entities
         public string SKU { get; set; }
         public string Description { get; set; }
         //relational
-        public int? CategoryId { get; set; }
+        public long? CategoryId { get; set; }
         public Category Categroy { get; set; }
         public ICollection<ProductTag> ProductTags { get; set; }
     }
