+++ OnionPronia/src/Core/OnionPronia.Domain/Entities/ProductTag.cs
@@ -10,7 +10,7 @@ namespace OnionPronia.Domain.Entities
     {
         public Product Product { get; set; }
         public Tag Tag { get; set; }
-        public int ProductId { get; set; }
-        public int TagId { get; set; }
+        public long ProductId { get; set; }
+        public long TagId { get; set; }
     }
 }
